refactor(box): share provisioning live reads
