+++ packages/drivers/box-provisioning/src/BoxProvisioningApplier.ts
@@ -7,7 +7,7 @@
 
 import * as B from "@beep/box";
 import { $BoxProvisioningId } from "@beep/identity";
-import { HttpsUrl, Sha256Hex } from "@beep/schema";
+import { Sha256Hex } from "@beep/schema";
 import { Context, DateTime, Effect, Equal, Layer, Match, MutableHashMap, MutableHashSet, pipe } from "effect";
 import * as A from "effect/Array";
 import * as O from "effect/Option";
@@ -29,12 +29,20 @@ import {
 import { BoxActionApplied, BoxActionBlocked, BoxActionSkipped, BoxApplyReceipt } from "./BoxProvisioningReceipt.ts";
 import {
   boxDesiredStateDigest,
-  canonicalObservedWebhook,
   canonicalWebhookIntent,
   digestText,
   encodedDigest,
   hasValidBoxProvisioningPlanDigest,
 } from "./internal/canonical.ts";
+import {
+  listFolderCollaborations,
+  listFolderItems,
+  listObservedWebhooks,
+  toObservedCollaboration,
+  toObservedCollaborationFromFull,
+  toObservedFolderFromFull,
+  toObservedWebhook,
+} from "./internal/live.ts";
 import type { BoxDesiredState, BoxLogicalKey } from "./BoxProvisioningIntent.ts";
 import type {
   BoxBlockedAction,
@@ -48,16 +56,9 @@ import type { BoxApplyOutcome } from "./BoxProvisioningReceipt.ts";
 
 const $I = $BoxProvisioningId.create("BoxProvisioningApplier");
 
-const markerPageLimit = 1000;
 const sha256Equivalence = S.toEquivalence(Sha256Hex);
 
 const RawResourceIdentity = S.Struct({ id: S.NonEmptyString });
-const RawWebhook = S.Struct({
-  id: S.NonEmptyString,
-  target: S.Struct({ id: S.NonEmptyString }),
-  address: HttpsUrl,
-  triggers: S.Array(S.NonEmptyString),
-});
 
 type DesiredResource =
   | BoxFolderIntent
@@ -66,39 +67,7 @@ type DesiredResource =
   | BoxMetadataIntent
   | BoxRetentionIntent;
 
-type MarkerPage<A> = {
-  readonly entries?: ReadonlyArray<A>;
-  readonly nextMarker?: string | null;
-};
-
 const invariant = (code: BoxProvisioningInvariantError["code"]) => BoxProvisioningInvariantError.make({ code });
-
-const collectMarkerPages = Effect.fn("BoxProvisioningApplier.collectMarkerPages")(function* <A, E>(
-  load: (marker: O.Option<string>) => Effect.Effect<MarkerPage<A>, E>
-): Effect.fn.Return<ReadonlyArray<A>, E> {
-  let marker = O.none<string>();
-  let entries = A.empty<A>();
-  const seen = MutableHashSet.empty<string>();
-
-  while (true) {
-    const page = yield* load(marker);
-    entries = A.appendAll(entries, O.getOrElse(O.fromNullishOr(page.entries), A.empty<A>));
-    const next = O.fromNullishOr(page.nextMarker);
-    if (O.isNone(next) || MutableHashSet.has(seen, next.value)) {
-      return entries;
-    }
-    MutableHashSet.add(seen, next.value);
-    marker = next;
-  }
-});
-
-const markerQuery = (marker: O.Option<string>) =>
-  O.match(marker, {
-    onNone: () => ({ limit: markerPageLimit }),
-    onSome: (value) => ({ limit: markerPageLimit, marker: value }),
-  });
-
-const markerFolderQuery = (marker: O.Option<string>) => ({ ...markerQuery(marker), usemarker: true });
 
 const findDesired = <T extends DesiredResource>(
   resources: ReadonlyArray<T>,
@@ -128,144 +97,6 @@ const resolveFolderProviderId = (
       onNone: () => Effect.fail(invariant("unresolved-dependency")),
       onSome: Effect.succeed,
     })
-  );
-
-type CollaborationPrincipal = {
-  readonly principal: string;
-  readonly principalProviderId: O.Option<BoxProviderId>;
-  readonly principalType: "user" | "group";
-};
-
-const makeCollaborationPrincipal =
-  (principalType: CollaborationPrincipal["principalType"], principalProviderId: O.Option<BoxProviderId>) =>
-  (principal: string): CollaborationPrincipal => ({ principal, principalProviderId, principalType });
-
-const collaborationPrincipal = (collaboration: B.Collaboration): O.Option<CollaborationPrincipal> => {
-  const pendingInvite = pipe(
-    O.fromNullishOr(collaboration.inviteEmail),
-    O.map(makeCollaborationPrincipal("user", O.none()))
-  );
-  return pipe(
-    O.fromNullishOr(collaboration.accessibleBy),
-    O.flatMap((accessibleBy): O.Option<CollaborationPrincipal> => {
-      const principalProviderId = pipe(O.fromNullishOr(accessibleBy.id), O.map(BoxProviderId.make));
-      if (S.is(B.UserCollaborations)(accessibleBy)) {
-        return pipe(
-          O.fromNullishOr(accessibleBy.login),
-          O.orElse(() => O.fromNullishOr(accessibleBy.id)),
-          O.map(makeCollaborationPrincipal("user", principalProviderId))
-        );
-      }
-      return pipe(O.fromNullishOr(accessibleBy.id), O.map(makeCollaborationPrincipal("group", principalProviderId)));
-    }),
-    O.orElse(() => pendingInvite)
-  );
-};
-
-const toObservedFolder = (folder: B.FolderFull): Effect.Effect<BoxObservedFolder, BoxProvisioningInvariantError> =>
-  O.match(
-    O.all({
-      name: O.fromNullishOr(folder.name),
-      parentProviderId: pipe(
-        O.fromNullishOr(folder.parent),
-        O.flatMap((parent) => O.fromNullishOr(parent.id))
-      ),
-    }),
-    {
-      onNone: () => Effect.fail(invariant("unreadable-sdk-response")),
-      onSome: ({ name, parentProviderId }) =>
-        Effect.succeed(
-          BoxObservedFolder.make({
-            etag: O.fromNullishOr(folder.etag),
-            name,
-            parentProviderId: O.some(BoxProviderId.make(parentProviderId)),
-            providerId: BoxProviderId.make(folder.id),
-          })
-        ),
-    }
-  );
-
-const toObservedCollaboration = (
-  collaboration: B.Collaboration
-): Effect.Effect<BoxObservedCollaboration, BoxProvisioningInvariantError> =>
-  O.match(
-    O.all({
-      folderProviderId: pipe(
-        O.fromNullishOr(collaboration.item),
-        O.flatMap((item) => O.fromNullishOr(item.id))
-      ),
-      principal: collaborationPrincipal(collaboration),
-      role: O.fromNullishOr(collaboration.role),
-    }),
-    {
-      onNone: () => Effect.fail(invariant("unreadable-sdk-response")),
-      onSome: ({ folderProviderId, principal, role }) =>
-        Effect.succeed(
-          BoxObservedCollaboration.make({
-            folderProviderId: BoxProviderId.make(folderProviderId),
-            principal: principal.principal,
-            principalProviderId: principal.principalProviderId,
-            principalType: principal.principalType,
-            providerId: BoxProviderId.make(collaboration.id),
-            role,
-          })
-        ),
-    }
-  );
-
-const toObservedWebhook = (webhook: unknown): Effect.Effect<BoxObservedWebhook, BoxProvisioningInvariantError> =>
-  S.decodeUnknownEffect(RawWebhook)(webhook).pipe(
-    Effect.map((decoded) =>
-      canonicalObservedWebhook(
-        BoxObservedWebhook.make({
-          address: decoded.address,
-          providerId: BoxProviderId.make(decoded.id),
-          targetProviderId: BoxProviderId.make(decoded.target.id),
-          triggers: decoded.triggers,
-        })
-      )
-    ),
-    Effect.mapError(() => invariant("unreadable-sdk-response"))
-  );
-
-const listFolderItems = (box: B.Box["Service"], folderId: BoxProviderId) =>
-  collectMarkerPages<B.Item, B.BoxError>((marker) =>
-    box.folders.getFolderItems(
-      B.FoldersGetFolderItemsPayload.make({
-        folderId,
-        optionalsInput: { queryParams: markerFolderQuery(marker) },
-      })
-    )
-  );
-
-const listFolderCollaborations = (box: B.Box["Service"], folderId: BoxProviderId) =>
-  collectMarkerPages<B.Collaboration, B.BoxError>((marker) =>
-    box.listCollaborations.getFolderCollaborations(
-      B.ListCollaborationsGetFolderCollaborationsPayload.make({
-        folderId,
-        optionalsInput: { queryParams: markerQuery(marker) },
-      })
-    )
-  );
-
-const listObservedWebhooks = (box: B.Box["Service"]) =>
-  collectMarkerPages<B.WebhookMini, B.BoxError>((marker) =>
-    box.webhooks.getWebhooks(B.WebhooksGetWebhooksPayload.make({ queryParams: markerQuery(marker) }))
-  ).pipe(
-    Effect.flatMap((webhooks) =>
-      Effect.forEach(
-        webhooks,
-        (webhook) =>
-          O.match(O.fromNullishOr(webhook.id), {
-            onNone: () => Effect.fail(invariant("unreadable-sdk-response")),
-            onSome: (webhookId) =>
-              box.webhooks
-                .getWebhookById(B.WebhooksGetWebhookByIdPayload.make({ webhookId }))
-                .pipe(Effect.flatMap(toObservedWebhook)),
-          }),
-        { concurrency: 1 }
-      )
-    )
   );
 
 const collaborationMatches = (
@@ -390,7 +221,7 @@ const validateFolderAbsent = Effect.fn("BoxProvisioningApplier.validateFolderAbs
   parentProviderId: BoxProviderId,
   desired: BoxFolderIntent
 ) {
-  const items = yield* listFolderItems(box, parentProviderId);
+  const items = yield* listFolderItems({ box, folderId: parentProviderId });
   const names = yield* Effect.forEach(
     A.filter(items, S.is(B.FolderMini)),
     (item) =>
@@ -410,8 +241,12 @@ const validateCollaborationAbsent = Effect.fn("BoxProvisioningApplier.validateCo
   folderProviderId: BoxProviderId,
   desired: BoxCollaborationIntent
 ) {
-  const collaborations = yield* listFolderCollaborations(box, folderProviderId).pipe(
-    Effect.flatMap((entries) => Effect.forEach(entries, toObservedCollaboration, { concurrency: 1 }))
+  const collaborations = yield* listFolderCollaborations({ box, folderId: folderProviderId }).pipe(
+    Effect.flatMap((entries) =>
+      Effect.forEach(entries, (collaboration) => toObservedCollaboration({ collaboration, folderProviderId }), {
+        concurrency: 1,
+      })
+    )
   );
   if (A.some(collaborations, (observed) => collaborationMatches(desired, observed, folderProviderId))) {
     return yield* invariant("action-precondition-mismatch");
@@ -443,7 +278,7 @@ const validateExistingPrecondition = Effect.fn("BoxProvisioningApplier.validateE
   const actual = yield* Match.value(action.resourceKind).pipe(
     Match.when("folder", () =>
       box.folders.getFolderById(B.FoldersGetFolderByIdPayload.make({ folderId: providerId })).pipe(
-        Effect.flatMap(toObservedFolder),
+        Effect.flatMap(toObservedFolderFromFull),
         Effect.map((folder) => ({ digest: encodedDigest(BoxObservedFolder, folder), etag: folder.etag }))
       )
     ),
@@ -451,7 +286,7 @@ const validateExistingPrecondition = Effect.fn("BoxProvisioningApplier.validateE
       box.userCollaborations
         .getCollaborationById(B.UserCollaborationsGetCollaborationByIdPayload.make({ collaborationId: providerId }))
         .pipe(
-          Effect.flatMap(toObservedCollaboration),
+          Effect.flatMap(toObservedCollaborationFromFull),
           Effect.map((collaboration) => ({
             digest: encodedDigest(BoxObservedCollaboration, collaboration),
             etag: O.none<string>(),
